Wrap ItemTable in server component and ErrorBoundary
